Reviewed details wrt free vs commercial
(full disclosure, I'm currently a Lightbend employee on the Lagom team)

Hi @AlexanderRay, I was reading through your docs and noticed some miss-conceptions wrt to what is free and what is commercial-only in the Lagom and ConductR offer.

Lagom has always been freely available for development and production. It is true that the recommended way to run Lagom in production is ConductR.
Just recently ConductR released a new licensing feature that makes it free to use in production with [up to 3 nodes](https://www.lightbend.com/platform/enterprise-suite) (registration required).